requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
fakeredis
python-dotenv
